requirements.txt
opencv-python
torch
torchvision
torchsummary
torchmetrics
torchtext
torchaudio
Jupyter
numpy
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-customvision
opencv-contrib-python
imutils
matplotlib
wand
scikit-learn
black
pillow==9.5.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
requests-toolbelt
nltk
jupyter
python-dotenv
openai
pydantic
scikit-image==0.20.0
image-similarity-measures
image-similarity-measures[speedups]
levenshtein
pyyaml
yacs
tqdm
lpips
streamlit
streamlit-option-menu
